refactor : AnswerBox에서 직접 editorVal 관리
