+++ frontend/components/community/ViewCommunity.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router';
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components';
 import { FaHeart } from 'react-icons/fa';
 import dynamic from 'next/dynamic';
@@ -28,28 +28,11 @@ const ViewCommunity = () => {
   });
   const data = communityQuery.data?.data;
 
-  //댓글 관련
-  const [answerVal, setCommentVal] = useState('');
-  const changeAnswerVal = (val: string) => {
-    setCommentVal(val);
-  };
-
-  const addAnswer = () => {
-    //댓글 추가 이벤트를 작성해주세요~~~
-  };
-
   //댓글 페이지 네이션
   const [answerPage, setCommentPage] = useState(1);
   const answerPageHandler = (num: number) => {
     setCommentPage(num);
   };
-
-  // //게시글에 해당하는 답글 데이터
-  // const { answerPageCount } = useGetAnswer({
-  //   category: 'ARTICLE',
-  //   postId: data?.articleId,
-  //   params: `size=5&page=${answerPage}`,
-  // });
 
   if (communityQuery.isLoading) return <Message>로딩중입니다.</Message>;
   if (communityQuery.error)
@@ -84,9 +67,6 @@ const ViewCommunity = () => {
               <AnswerBox
                 answerPage={answerPage}
                 answerPageHandler={answerPageHandler}
-                addAnswer={addAnswer}
-                answerVal={answerVal}
-                changeAnswerVal={changeAnswerVal}
               />
             </div>
           </Bottom>
